Remove obsolete CoreMainMenuController
