+++ MonoChrome/Assets/Scripts/README_PORTFOLIO.md
@@ -14,6 +14,9 @@
 - **싱글톤 패턴** (메모리 누수 방지)
 - **의존성 주입** (Dependency Injection)
 
+> **Note**: 기존 `GameManager`와 `CoreGameManager` 클래스는 `MasterGameManager`로 완전히 통합되었습니다.
+> 레거시 호환을 위해 존재하지만 새 코드에서는 사용을 지양합니다.
+
 ---
 
 ## 🏆 포트폴리오 어필 포인트
@@ -24,7 +27,7 @@
 | 높은 결합도 (직접 참조) | 낮은 결합도 (이벤트 통신) | **확장성 ↑** |
 | 낮은 응집도 (혼재된 책임) | 높은 응집도 (단일 책임) | **유지보수성 ↑** |
 | 테스트 불가능 | 단위 테스트 가능 | **품질 보장** |
-| GameManager 500+ 줄 | 200줄로 단순화 | **60% 코드 감소** |
+| GameManager 500+ 줄 (현재는 MasterGameManager로 통합) | 200줄로 단순화 | **60% 코드 감소** |
 
 ### ✅ 실무 역량 증명
 - **리팩토링 경험**: 기존 코드의 문제점 파악 → 체계적 개선
@@ -101,7 +104,7 @@ public class UIController : MonoBehaviour
 1. **핵심 시스템 생성**:
    ```
    Hierarchy:
-   ├── [ImprovedGameManager] (DontDestroyOnLoad)
+   ├── [MasterGameManager] (DontDestroyOnLoad)
    ├── [GameStateMachine] (DontDestroyOnLoad)
    ├── [EventBus] (DontDestroyOnLoad)
    ├── [DungeonController]
