+++ MonoChrome/Assets/Scripts/Setup/ProjectSetupAndValidator.cs
@@ -108,12 +108,12 @@ namespace MonoChrome.Setup
         {
             Debug.Log("1. 핵심 컴포넌트 생성 중...");
 
-            // CoreGameManager
-            if (CoreGameManager.Instance == null)
+            // MasterGameManager
+            if (MasterGameManager.Instance == null)
             {
-                GameObject gameManagerGO = new GameObject("[CoreGameManager]");
-                gameManagerGO.AddComponent<CoreGameManager>();
-                Debug.Log("✓ CoreGameManager 생성됨");
+                GameObject gameManagerGO = new GameObject("[MasterGameManager]");
+                gameManagerGO.AddComponent<MasterGameManager>();
+                Debug.Log("✓ MasterGameManager 생성됨");
             }
 
             // GameStateMachine
@@ -185,7 +185,7 @@ namespace MonoChrome.Setup
         {
             Debug.Log("1. 핵심 시스템 존재 검증 중...");
 
-            CheckComponent<CoreGameManager>("CoreGameManager");
+            CheckComponent<MasterGameManager>("MasterGameManager");
             CheckComponent<GameStateMachine>("GameStateMachine");
             CheckComponent<EventBus>("EventBus");
             CheckComponent<DungeonController>("DungeonController");
