+++ MonoChrome/Assets/Scripts/Systems/UI/MainMenuController.cs.meta
@@ -1,2 +1,2 @@
 fileFormatVersion: 2
-guid: 084e61cec76e30149b592c6410222162
+guid: cc22d3b7-5f29-4131-9310-f5d485dd1554
